requirements.txt
requests==2.21.0
python-dotenv==0.10.1
terminaltables==3.1.0